rename pages to views and fix paths
