+++ src/client/src/react/components/MainScore.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "./MainScore.css";
 
-import PlayerScore from "../PlayerScore/PlayerScore";
+import PlayerScore from "./PlayerScore";
 import { setsWonPlayer1, setsWonPlayer2 } from "../lib";
 
 function MainScore({ match, player1, player2 }) {
