+++ src/client/src/react/components/ScoreBoard.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "./ScoreBoard.css";
 
-import MainScore from "../MainScore/MainScore";
+import MainScore from "./MainScore";
 import SetScore from "./SetScore";
 
 function ScoreBoard({ match, player1, player2 }) {
