+++ src/client/src/react/views/LoginView.js
@@ -1,23 +1,23 @@
 import React from "react";
-import "./Login.css";
+import "./LoginView.css";
 
-function Login({
+function LoginView({
   tableNumber,
   sendTableNumber,
   tableNumberChanged,
   availableTables
 }) {
   return (
-    <div>
+    <>
       <div className="login">
         <form className="login-form">
-          <datalist id="table-numbers">
+          {/* <datalist id="table-numbers">
             <option value="0" />
             <option value="1" />
             <option value="2" />
             <option value="3" />
             <option value="4" />
-          </datalist>
+          </datalist> */}
           <label htmlFor="login-input-field" className="login-text">
             Tischnummer
           </label>
@@ -40,8 +40,8 @@ function Login({
           </button>
         </form>
       </div>
-    </div>
+    </>
   );
 }
 
-export default Login;
+export default LoginView;
