+++ src/client/src/react/views/WaitingView.css
